remove ParseAST code from protocol and driver
Initially we thought we would expose the raw native AST
to the external interface. This is not the case, since
all requests going in and out of the driver use *uast.Node
type, so the only difference is if it's annotated or not.

This commit removes ParseAST from the protocol and leaves
just ParseUAST. This reduces code paths, since parsers at
different levels can all use the same interface (UASTParser).

ASTParser still exists but only in the 'native' package, as
an internal detail about communciation betwee native parser
and UAST conversor inside a driver.