transformer: Any op should mean "any node"; rename old one to AnyElem
This commit introduces a minor breaking change. Previously "Any" op
was used to check that "any node in array matches an argument", while
it's more natural for this name to mean "any node". The old op was
renamed to "AnyElem", while a new "Any" helper was defined as an alias
for AnyNode(nil).

This commit also clarifies the documentation on those operators and
moves array-related operators to an appropriate file.

Signed-off-by: Denys Smirnov <denys@sourced.tech>