driver: do not force roles for native uast files
Due to the way how uast.RolesOf is implemented, it will always addend
Unannotated roles to all nodes that do not define roles and are not part
of the Semantic UAST schema.

YAML encoder that is used to write fixtures in tests uses this function
to dump native AST. As a result, all AST nodes end up with a Unannotated
tole and a new @role field. This breaks the transformation tests because
of the unknown field that should not exist for Native parsing mode.

This commit adds an option for the YAML UAST encoder to disable roles
for native AST. Test are changed according, so drivers will need to
re-generate native fixtures after the next SDK update.

Signed-off-by: Denys Smirnov <denys@sourced.tech>