transformer: optimize field descriptors
Field descriptors are currently represented with a Go map. But it turned
out that iteration over those maps take a significant part of the DSL
processing time. This is mostly because of extensive use of Part
transform that cannot be optimized by the DSL.

This commit changes the field descriptors to a struct that contains a
slice of fields. After the change the processing time decreases by 30%.

Technically, this is a breaking change, but it will only affect drivers
that define custom ObjectOps.

Signed-off-by: Denys Smirnov <denys@sourced.tech>