build: use Go scripts instead of bblfsh-sdk binary; resolves #325
There are quite a few problems with building using bblfsh-sdk tool.
The most significant one is that bblfsh-sdk binary may be built with a
different SDK version, than the driver expects. It will cause the tool
to update driver files, while the driver will fail to build using those
files either locally, or in CI where the correct tool version will be
available.

Since we have a copy of SDK in the driver's vendor directory anyway,
use it to build the driver. This is done by two small Go scripts in the
driver directory. They will call the build or test command the same way
as the bblfsh-sdk tool does.

The change does not affect the bblfsh-sdk tool itself - it can still be
used to initialize new drivers or update existing ones.

Signed-off-by: Denys Smirnov <denys@sourced.tech>