protocol: set a default max message size for all clients and servers
Currently clients call DialOptions to get the default dial options that
should work well for a Babelfish gRPC client. But this config does not
include the default size setting that is set for the server.

This change sets the default message size in both the client and the
server options.

Signed-off-by: Denys Smirnov <denys@sourced.tech>